backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
openai==1.10.0
networkx==3.2.1
numpy==1.26.3
pandas==2.2.0
reportlab==4.0.8
openpyxl==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.13.1
pytest==7.4.4
pytest-asyncio==0.23.3
aiohttp==3.9.1
